Added tests for compiler's --print-callstack-on-error.
Note 1. The exact format is of course insignificant. The point is that
the stack be present.

Note 2. The compiler uses some heuristics to squash duplicate error
messages. Ideally, with --print-callstack-on-error, it would squash
duplicates only if their stack traces match.  Currently it is more
aggressive than that, but we leave the way it is for now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18031 3a8e244f-b0f2-452b-bcba-4c88e055c3ca