Replaced INT_ASSERT with code to handle the formerly-failing case.
This fixes the ill-formed class method bug reported by Christopher Swenson,
namely the compiler failed with an assertion on this:

  record A {
    def A.a() {}
  }

Chapel spec allows prefixing a class method with a type binding ('A.' here),
so I added code to handle this case.  (I also added a couple of test cases.)

For now, I only allow the type binding to refer to the innermost class
(via checking the type's name against the class's name), to simplify the fix.
Maybe we want to impose this restriction in the language?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18036 3a8e244f-b0f2-452b-bcba-4c88e055c3ca