Added a warning for apparent misunderstanding of 'use' statements
A common problem users run into is to think of use statements like
#includes and write code like the following:

          foo.chpl
          ------------
          use A, B, C;
          use D;

          module bar {
          }
          ------------

When this really has the effect of creating a top-level module named
foo based on the file with only use statements containing a submodule
named bar.  Since use statements are not #includes but code, this
interpretation is consistent and correct (and may even *occasionally*
be what you wanted), but is more likely to be a thinko and one that is
hard to debug.  This came up this week in a mail from a user but is
not the first time we've seen it.

To this end, we now generate a warning along the following lines:

foo.chpl:4: warning: as written, 'bar' is a sub-module of the module created for file 'foo.chpl' due to the file-level 'use' statements.  If you meant for 'bar' to be a top-level module, move the 'use' statements into its scope.

Fixed two tests that ran into this warning -- one in the spec, one not
-- but worked fine anyway based on their simplicity.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18043 3a8e244f-b0f2-452b-bcba-4c88e055c3ca