Implementing an eof check for files
Based on comments from Michael Ferguson and Chris Swenson, adding an
eof check to the file type.  This implementation is slightly fragile
in that it assumes C's feof() call returns Chapel's int type (or is
compatible with it).  The better solution here would be to have a
CTypes.chpl module that defined a c_int type that could be used to
declare the return type of feof(), but I wanted to get something that
worked in sooner than waiting for the implementation of this alias...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18051 3a8e244f-b0f2-452b-bcba-4c88e055c3ca