Backed out r18036 fix (ill-formed class method bug from Christopher Swenson) because it is not robust. Instead made it an error saying "currently we don't handle this". Here is a code example:
  record A {
    def A.a() {}  // defines a() on A
  }

The test 'currently-error.chpl' verifies that the compiler
currently handles this case, generating an error.

However, we discussed this in the group meeting on 2010-12-09 and
agreed that ideally we would like to interpret the idendifier in the
"type binding" ('A.' here) via usual name resolution. (Name resolution
should happen early enough for the compiler to decide what type to
attribute the method ('a' here) to.)

Therefore we also want to allow type bindings where the type
is not the immediately-enclosing class, e.g.

  record A {
    def int.intfun() {}  // defines intfun() on int
  }

There was a general consensus that the visibility of a thusly-
defined method would be restricted to the lexical scope
of its definition. E.g. the above 'intfun' would not be
legal outside of the definition of record A. It's like
declarations in a module are not visible outside the module
(unless that module is "use"-ed).

We might want to double-check it, however:

- We get different visibility depending on whether the type
binding-qualified method in question belongs on the
immediately-enclosing class or something else.

- Does the visibility scope include methods on the enclosing
class syntactically defined outside the class, e.g.

  record A {
    def int.intfun() {}  // defines intfun() on int
  }
  def A.b() { 1.intfun(); } // legal?


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18052 3a8e244f-b0f2-452b-bcba-4c88e055c3ca