Disclaimer: Greg gets all the credit for finding 'stty_init' and finding the presently-working solution as follows:
(a) Expect has this variable, 'stty_init', which affects "spawn"-created ptys.
    So if I set it to -onlcr, I won't get CRs. E.g.

        set stty_init -onlcr
        spawn -noecho echo no ^Ms here
        interact

(b) That does not work for us. That's because 'qsub -I' creates its own pty,
    which is unknown to expect.

(c) So we need to mess with qsub's pty. Namely:

        spawn "qsub -I ....."
        send "stty -onlcr; aprun ....."
        interact

(c') Turns out that if we instead do

        spawn "qsub -I ....."
        send "stty -onlcr; set prompt ....."
        expect ...
        send "aprun ....."
        interact

    then it doesn't work. (Greg, can you figure out this one? :)

(d) However, the above squashes CRs even when run in a terminal proper
    (so the output looks funny).

(e) So we need to check whether stdout is a terminal or not -- and
    then add 'stty -onlcr' or not -- in the launcher:

         fprintf(expectFile, "send \"%s aprun .....",
                 isatty(fileno(stdout)) ? "" : "stty -onlcr;" ....);

This change let me run successfully start_test on the examples directory
(with mods to start_test/sub_test to be checked in later).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18071 3a8e244f-b0f2-452b-bcba-4c88e055c3ca