Added '-launchertimeout pbs' option to start_test/sub_test and also implemented '-nostdinredirect' if !useTimedExec.
The -launchertimeout option relies on the launcher's timeout facility
(in the case of PBS, the "walltime" resource), instead of running the
compiled executable under timedexec. This is useful on XT/XE where
running the executable under timedexec makes it hang (for an unknown
reason). Note that if we take out 'eval' from timedexec then it works
fine, but we have not convinced outselves that timedexec would still
reliably kill the process.

Right now -launchertimeout takes an additional argument indicating
which launching system is being used (e.g. 'pbs') so that start_test
can pass an appropriate timeout option to the launcher. However, given
that currently we have only one launcher, this is not necessary and we
can take it out.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18072 3a8e244f-b0f2-452b-bcba-4c88e055c3ca