Fixing an XMT regression I added with r18055
When doing the tasking refactoring of late fall 2010, I added a
warning to the XMT layer if the maxThreadsPerLocale value was set to
something other than 0 on the belief that I was going to have it
always be initialized to the sentinel value of 0 by default.  But then
I never made that change, so the warning printed out on every
execution.

Turns out that the value of maxThreadsPerLocale is never used on an
XMT which makes setting it to 100*coresPerLocale (whatever value that
is) turns out to be pointless, so in this change I make the two
maxThreads calls return 0 making the warning meaningful.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18087 3a8e244f-b0f2-452b-bcba-4c88e055c3ca