Improved ordering of initialization code for tasking layer
Prior to this change, our internal ChapelThreads.chpl modules would be
initialized to set up the maxThreadsPerLocale and callStackSize
configs prior to calling CHPL_TASKING_INIT().  As part of this, it
would query the maximum callstack size limit from the tasking layer.
However, that is weird because it calls into the tasking layer prior
to initializing it, which is arguably weird (and caused some problems
with the qthreads implementation which wants to make a qthreads query
prior to initializing the qthreads library).

This fix changes that, and generally simplifies the tasking interface,
by putting the check against any task-layer-specific limit into the
CHPL_TASKING_INIT() call itself rather than passing information from
the runtime to the internal modules for the check only to pass the
callStackSize value back down to the INIT() call after the fact.  The
only slight downside to this approach is that it pushes a Chapel
warning call into the runtime where it is somewhat less flexible/easy
to use compared to within the Chapel code.  The approach I took also
results in a bit of duplicated code w.r.t. the generation of this
warning, although arguably no more than already exists between the
multiple files that make calls to getrlimit() and setrlimit().

Here are the changes by file:

Removed query of limit and check against it from internal modules:
       modules/internal/ChapelThreads.chpl

Removed primitive from compiler:
       compiler/include/primitive.h
       compiler/AST/primitive.cpp
       compiler/AST/expr.cpp

Removed routine from tasking interface:
       runtime/include/chpltasks_func_names.h
       runtime/include/chpltasks.h
       runtime/src/tasks/mta/tasks-mta.c

Moved code related to checking the limit into the initialization
function and issue a warning if we exceed it:
       runtime/src/tasks/none/tasks-none.c
       runtime/src/tasks/fifo/tasks-fifo.c
       runtime/src/threads/pthreads/threads-pthreads.c




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18088 3a8e244f-b0f2-452b-bcba-4c88e055c3ca