Added '_extern' as a keyword, so emacs can color it properly.
I did not investigate exactly what categories it should be added to,
so I added it to all the categories that "def" occurs in.
If that doesn't work in some case(s), we will need to investigate.

We could add the same to chpl-mode for earlier versions of emacs, too.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18095 3a8e244f-b0f2-452b-bcba-4c88e055c3ca