A start at a "best practices" notes repository
I got a late start on this this week (who wants to write when there
aren't regression tests being run?), but got a few notes typed up this
evening to get a sense of what I thought we might want.  I've also
started sorting the other things we wanted to cover into rough
categories and will try to go through them in tomorrow's meeting,
possibly looking for volunteers to help.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18099 3a8e244f-b0f2-452b-bcba-4c88e055c3ca