made CHPL_TASKING_CALL_MAIN() standard/reordered initialization of runtime
This change makes all tasking layers support CHPL_TASKING_CALL_MAIN()
as the means of kicking off the user's code via the chpl_main() call.
In particular, the tasking layer is passed a pointer to chpl_main()
and it can choose to simply call it directly, or to create a task and
have that task call it.  This removes the RUN_MAIN_AS_A_TASK macro and
the #if... conditionals that checked for it.

This change also moves the call to chpl_init_chpl_rt_utils() (which
initializes runtime/src/common-chpl/chpl_rt_utils.chpl to occur after
CHPL_TASKING_INIT().  The reason for this is that that Chapel code
contains sync vars and we shouldn't assume that a tasking layer can
support CHPL_SYNC calls before it is initialized (in particular,
qthreads cannot).

The reason these two changes are related is that the fifo tasking
layer used to insert the main task into the task table as part of its
INIT() routine, but this resulted in a cycle since the task table
isn't ready to use until after chpl_rt_utils.chpl has been
initialized.  Greg and I decided that it would be better to make
CHPL_TASKING_INIT() take care of all general initialization for the
tasking layer and then to have CHPL_TASKING_CALL_MAIN() do things
related to setting up the main task (including adding itself to the
task table).  To that end, a fair amount of code moved from
CHPL_TASKING_INIT to CHPL_TASKING_CALL_MAIN() in the fifo tasking
layer.

In the mta and none tasking layers, CHPL_TASKING_CALL_MAIN() simply
calls chpl_main().  It didn't seem like anything else in
CHPL_TASKING_INIT() related to establishing the main task.

We should now be at a point where no sync vars are used before
CHPL_TASKING_INIT() has been called.  With any luck, this should
permit us to run hello-multicore.chpl with qthreads now.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18102 3a8e244f-b0f2-452b-bcba-4c88e055c3ca