A couple of trivial changes to enums
This is probably my first five minute change in years.  Removed the
required semicolon from the end of the enumerated type declaration;
and permitted the final enumerated symbol to be followed by a comma.
Documented the changes.  Please give a shout if you disagree (nobody
did on the thread in time for me to hold off on this once testing
finished :).

Added a feature request future for being able to 'use' an enumerated
type to inject its symbols into the current scope.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18108 3a8e244f-b0f2-452b-bcba-4c88e055c3ca