A bug in how the display of warnings for compilerWarning() calls is affected by the 'errorDepth' argument (when not 1).
The .future also contains a description of another bug.
I could not put together a clean reproduction of that bug easily,
so I am describing how to reproduce it in prose.
So when someone is addressing the but exhibited by repeated-warning-1.chpl
per se, maybe he/she could look at this other bug as well.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18127 3a8e244f-b0f2-452b-bcba-4c88e055c3ca