Starting work on a new revision of HPL
To ramp up on HPL in 2011, I'm taking a pass over HPL to try and
clean it up w.r.t. unnecessary assertions and other clean-ups
that we noticed (but didn't fix) when we last looked at it in
May.  This change:

(1) adds a new version of HPL that I'm going to make my edits in until
I think they're good enough to collapse back into the release and

(2) changes the blocksize of the original hpl example so that it's not
a multiple of the problem size (which exercises some more interesting
code paths and increases my confidence that my changes are correct).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18129 3a8e244f-b0f2-452b-bcba-4c88e055c3ca