Revert to "module access without use" behavior
This commit takes us back to the pre-r15529 behavior where a module's
symbols can be accessed without first using that module.  In
particular, you can now do:

            module M1 {
              var x: real;
            }

            module M2 {
              ...M1.x...
            }

whereas since r15529, we have required the presence of 'use M1' to be
in a containing lexical scope of the expression 'M1.x'.  The fix
itself is quite simple and makes the compiler simpler (as measured by
lines of dead code removed).  Most of the changes here are additions
of new tests to lock in this behavior, updating of old tests that
generated error messages due to the lack of a 'use', and removing now
unnecessary 'use' statemets that were introduced in r15529.

What changed?  At the time of r15529, module initialization was still
triggered by the first dynamically encountered 'use' of a module.
This led to some concerns for expressions like M1.x -- would every
such expression need to be guarded by code to ensure that M1 had been
initialized?  There were also complications related to function
resolution order when trying to resolve functions in a module that may
not yet have been (dynamically) initialized.

Now that module initialization is done completely at program startup
time, I think the original concerns are no longer warranted.
Moreover, making this change didn't cause any existing tests to fail
and permitted a number that had been complicated by 'use' statements
to look simpler/more like their original form (meaning we've taken a
net step forward since not all of these tests were working pre-15529.

My rationale for why this makes sense is that "M1" is in scope for
expressions like "M1.x" if module M1 has been parsed (which is the
same reason that it can be 'use'd in the first place), and being able
to refer to symbols within a module scope in this way seems very
consistent with enum references, class/record member references, etc.

In discussing this (and part of the original motivation for giving it
a try), we discussed the fact that 'use' currently means "find this
module and parse it" as well as "import its symbols".  To that end, we
discussed the desire to eventually decouple the first behavior from
the second, either using a new keyword ("load M1" was proposed by
Michael Ferguson) or by using a degenerate form of the (planned but
yet-to-be-specified) module filtering operator (e.g., "use M1 only /*
nothing */;").




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18130 3a8e244f-b0f2-452b-bcba-4c88e055c3ca