Filing a pair of bugs related to externs and empty records
In the compiler, we currently prune empty records fairly aggressively
from the AST.  For code that is completely owned by Chapel this is
reasonable, but in implementing support for extern functions and
records, we did not dial back the aggression.  The result is that
calls made to extern functions are not made properly if it is passed
an external record (whether internal or external).

Moreover, we should never prune apparently external empty record
variables with no fields because we permit users to specify as many or
few of the fields of external records as they wish.

These tests capture two cases of this bug, one using a Chapel record,
one using an external record type.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18138 3a8e244f-b0f2-452b-bcba-4c88e055c3ca