Adding a best practices notes that I owed on the testing system
This is a brief introduction to the testing system that points out
some highlights and the philosophy and then refers to start_test's
comments for all the details.  Longer-term, I think we should move
the documentation out of start_test and into this file.

I moved the description of the different .future types to this 
file rather than the old README.futures file (which was out of
date and which I could never remember where it was when trying
to point people to it).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18142 3a8e244f-b0f2-452b-bcba-4c88e055c3ca