This commit contains partial progress towards the conversion from 'def' to 'proc' and 'iter' keywords.
Complete spec will appear in a subsequent commit message.
At this point, the following is done:

- added 'proc' and 'iter' keywords
- enabled 'return' with no expression in iterators
- added a warning upon a 'return' with an expression in a 'def' iterator
  (i.e. iterator declared with the 'def' keyword)
- added an error upon a 'return' with an expression in an 'iter' iterator
- added the options (presently unused):
   --enforce-proc-iter-keywords - enforce the new rules (with errors)
   --print-proc-iter-conversion-messages - print conversion messages
     to be used by the code conversion tool (I might not need this one)

Added tests for the above changes:
  test/functions/vass/proc-iter/

For now, only the implemented warnings/errors are reflected in:
  conversion-messages.good

Modified existing tests for compiler changes:
 - replaced the identifier 'iter' with 'itr'
  test/functions/iterators/diten/side_effect_iterator.good
  test/functions/iterators/diten/iteratorNoYield.chpl
  test/functions/iterators/diten/iteratorNoYield.good
  test/functions/iterators/diten/side_effect_iterator.chpl
 - replaced the identifier 'iter' with 'iters' or some such:
  test/arrays/deitz/jacobi5-no-local.chpl
  test/arrays/deitz/jacobi5.chpl
  test/arrays/deitz/jacobi5a.chpl
  test/distributions/robust/arithmetic/reindexing/test_reindexing4.chpl
  test/npb/ft/ft-serial.chpl
  test/studies/shootout/mandelbrot/mandelbrot-complex.chpl
  test/studies/shootout/mandelbrot/mandelbrot-dist.chpl
  test/studies/shootout/mandelbrot/mandelbrot-fancy.chpl
  test/studies/sudoku/dinan/sudoku-simple.chpl
 - replaced 'return expr' with yield+return
  test/functions/deitz/iterators/test_instantiate_iterator2a.chpl
  test/functions/deitz/iterators/test_mutually_recursive_iterators.chpl
  test/functions/iterators/bradc/iteratorBadError.chpl
  test/functions/iterators/bradc/iteratorBadError.good


Implementation notes.

check_functions() in semanticChecks.cpp: refactored (e.g. it does not
create a vector of returns) and strengthened (added checks for yields
and returns with expressions). Also removed a call to typeInfo(),
which might be problematic, e.g. if you compile
  if ... then return __primitive("complex_get_real", 11.2); else return;
(Note there is another call to typeInfo() in functionResolution.cpp
which will be problematic for this code.)

isIterator: In a couple of places I inserted a check for isIterator
(e.g. 'isIterator && is_void_return(ret)') to bypass a subsequent
check in certain cases, even though it would be correct without isIterator
as well.

Refactored some code in normalize_returns() to separate out the
conversions a return or yield statement undergoes.

Added the flag FLAG_PROC_ITER_KW_USED to indicate FnSymbols
declared with 'proc'/'iter' rather than 'def'.
After the transition period, all FnSymbols will have this flag set,
so we can retire the flag and all its uses.

Coincidentally:

Replaced USR_FATAL() with USR_FATAL_CONT() in many cases so it can
report multiple errors more often (mostly in semanticChecks.cpp).
This is not important, but I updated correspondingly:

Changed the error text "only allowed at module scope" to
"are allowed only at module scope".
Correspondingly, updated
  test/trivial/bradc/localConfig.good

Changed the error text "Not all function returns return a value"
to "Not all returns in this function return a value".
Correspondingly, updated
  test/functions/deitz/test_return2.good

Allowing 'return' with no expression in an iterator indirectly
led to a bug where the compiler went into an infinite recursion
on mistakenly-written code. Fixed by adding USR_STOP() after
resolveFns(chpl_main) in resolve(). Added such code as a test case:
  test/functions/vass/passing-iterator-as-argument.chpl

Improved the test message when the return type of an iterator
is declared as 'void' (exactly 'void', rather than something
that resolves to 'void' - for the latter, there is a separate future).
Added a test:
  test/functions/iterators/vass/direct-void-iterator-error.chpl
  test/functions/iterators/vass/direct-void-iterator-error.good

Cleaned up a couple of option descriptors in arg_desc[].

Added a function just like collectCallExprs() but that collects
expressions directly in a given function, not in any nested function:
  collectMyCallExprs(ast, vec, fn)


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18181 3a8e244f-b0f2-452b-bcba-4c88e055c3ca