Fixed a bug related to extern empty records
In general, the Chapel compiler removes empty records from the generated
code as a performance optimization/way to avoid warnings in the C code.
In the case of external records, we permit any subset (including none)
of the fields to be declared within the Chapel code, only permitting
those that are referenced to be accessed within Chapel.  An impact of
this was that such records were getting eliminated from the Chapel
compiler's code even though (a) the Chapel compiler did not own the type
in question, and (b) the record may not have been empty in practice.

This is a simple fix which simply removes external records from 
consideration as empty records are being removed.

I did not fix the somewhat more subtle problem of passing an empty Chapel
record to an external function, in part because it's a more subtle
problem (all uses of the record must be examined) and in part because
it's not entirely clear what the motivation for doing this would be.
A future still exists for that case.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18182 3a8e244f-b0f2-452b-bcba-4c88e055c3ca