Fix bug with extern functions that return records
This commit fixes a bug regarding external functions that return
records.  For Chapel code, functions that return records are currently
converted into functions that pass the result record back and forth
through an argument for purposes of memory management.  This
transformation was firing on Chapel's stubs for external functions as
well, causing the generated code to try and pass the result record in
as an argument rather than capture it as a return value.  This is
clearly not permissable since Chapel does not own the external code.
This fix disables the transformation for external functions and files
a test to lock that behavior in.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18184 3a8e244f-b0f2-452b-bcba-4c88e055c3ca