Initial support for external "classes"; disabling of extern subclasses/records
This commit adds initial support for external "classes" in C code.  As
C does not support classes, an external class is expected to be a
pointer to a struct.  This change requires that for a declaration of
the form _extern class C { ... }, the pointer-to-struct should be
typedef'd using the name C and the struct itself should be typedef'd
to have the name _C.  It may be possible to relax these assumptions if
necessary, but for now they simplified the implementation since it
matches what the Chapel compiler does for its internal implementation
of classes.  Like records, the member fields of the class may be
partially or completely specified.

This commit also disables the ability to declare external classes and
records that are subclasses/records of another class (internal or
external) because it wasn't clear to me in the time I've thought about
it what this should mean or how it would work.  Again, this can be
revisited if necessary down the road.

Part of making this change was to convert the parser such that in the
case of no superclass it returns NULL rather than an empty call to a
primitive as in previous commits.  While build.cpp seemed designed to
handle this, I completely expected something else in the code to
break, but under testing nothing did.  This led me to wonder if there
were other parts of the parser that used a similar idiom.  In
particular, looking back through the history of this change, it seems
that we used to return an empty AList() (which was our placeholder for
any AST rather than using NULL pointers), then converted it to use a
CallExpr as part of a global change to use AST nodes everywhere for
the purposes of our internal garbage collection.  It seems that this
decision was never revisited in order to see whether a plain-old NULL
pointer could be used.  Should be a modest space/time savings for
compilation in the long-run for any code that uses root
classes/records (which is most everything given our internal use of
these types).

I did not add documentation to README.extern for these features for
the time being because I wanted to live with them (and have users kick
their tires a bit) in order to raise confidence that they were useful
as implemented and somewhat stable.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18185 3a8e244f-b0f2-452b-bcba-4c88e055c3ca