This commit makes all remote get operations use the CHPL_COMM_GET() macro, which may use non-blocking communication if supported by the communication and tasking layers.  This required explicit passing of the locale and the remote address (rather than the wide reference), since the remote address is computed differently depending on the type of get operation is being performed.  Also, in preparation for bulk transfers, the type argument was replaced with a size argument.
In addition, the comm layer's blocking get operation is now exported
to the tasking layer via the macro CHPL_COMM_BLOCKING_GET().  This
allows the tasking layer to use the blocking get if it is more
appropriate (e.g., based on message size).  This resulted in the
renaming of comm_nb_macros.h to comm_task_decls.h since the it seemed
a more appropriate name.

NOTE: While I believe this works with the PVM multirealm environment,
it's still a bit of a hack.  A forthcoming commit with change the
interface the the chpl_comm_put/get to take both a type and a length.

A few other changes:

- Fixed a bug that was caused by a macro-local variable name
  conflicting with a user variable name

- Removed the definition of CHPL_COMM_BLOCKING_GET() in chplcomm.h,
  since all get operations user CHPL_COMM_GET() now

- Realigned the lovingly placed '\' in macros




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18191 3a8e244f-b0f2-452b-bcba-4c88e055c3ca