This change removes chpl_begin_cnt and chpl_can_exit from MTA tasking.
This change removes chpl_begin_cnt and chpl_can_exit from MTA tasking. These
are sync variables, and they are redundant given the code that calls the
MTA tasking layer keeps track of when to exit. This was tested with a program
used to test this same change with qthreads where the child task outlives
the parent task. (This change also works with STREAM, though no performance
improvement was obvious.)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18192 3a8e244f-b0f2-452b-bcba-4c88e055c3ca