This change adds a new child-parent test.
This change adds a new child-parent test. This runs quickly (in half a second)
on my desktop. It is designed to have the parent exit well before the child
exits. This test was written by Brad, and it was written to confirm the
removal of a sync variable would work with qthreads (later used to prove the
same thing on the XMT).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18193 3a8e244f-b0f2-452b-bcba-4c88e055c3ca