Making use of -Wmissing-declarations more sensitive to gcc version number
It turns out that -Wmissing-declarations is only supported by g++ as
of version 4.3.x which broke most of our nightly builds since the
default gcc module is 4.2.x.  Since this is a fairly recent change
in g++, I made the Makefile.gnu file query the g++ version number
and apply the warning to either the C++ or C warning flags list
based on whether or not the version is greater than 4.2.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18199 3a8e244f-b0f2-452b-bcba-4c88e055c3ca