This commit propagates Vass's proc/iter changes to the spec. I also noticed that some module still used def instead of proc or iter as appropriate, so I updated those as well.
The compiler issues an error if a function is declared as an iterator 
and does not contain a yield statement.  So I had to update some trivial
iterators by adding 
	yield 0;
to get them to compile.  Otherwise, the only changes were the def->proc and def->iter
syntax changes.

The changes in the bestPractices documents are just typos or grammatical errors I came across.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18200 3a8e244f-b0f2-452b-bcba-4c88e055c3ca