Fixing a bug in providing paths to file constructors
Since February 2008, it seems that the "path" argument to a file
constructor has been summarily ignored when opening files (making it
pretty much worthless).  This commit fixes the bug and adds a test to
make sure it doesn't break again.

That said, note that I don't particularly find the current file
interface to be particularly ideal.  This is more a maintenance of the
draft world we implemented and have never improved upon rather than an
assertion that we're on the right path.

Added a comment to the start_test documentation describing the purpose
of CLEANFILES which seems to have never been documented (or it got
lost?)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18213 3a8e244f-b0f2-452b-bcba-4c88e055c3ca