Attempt to get intel build of compiler working again
It looks like our intel testing is working better again but that the C
code generated for chapel.ypp is tripping over a new warning/error
(not our fault).  Throwing a flag to try and squash it with the caveat
that it's currently not easy enough to load an intel compiler to test
that this fixes the problem, so I'm relying on the nightly testing.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18217 3a8e244f-b0f2-452b-bcba-4c88e055c3ca