Added a script to perform semi-automatic source code conversion for the def->proc/iter change in the Chapel language. Changes in the compiler needed for this script to operate will be checked in soon.
(From convert-defs -h:)

This script is provided so the Chapel users can semi-automatically
adjust their code to certain changes in the Chapel language presently
being implemented. A brief summary of changes is:
- instead of the 'def' keyword, use:
   - 'proc' for non-iterator functions,
   - 'iter' for iterators;
- correspondingly, 'proc' and 'iter' are now keywords and may not
  be used as identifiers
  ('def' will not be a keyword after the transition is completed);
- iterators may not 'return' an expression -
  a 'yield EXPR; return;' sequence must be used instead.

This script attempts to convert Chapel source files by running the
compiler on each file and looking at the "deprecated" warnings that
result. Or, file/lines for def->proc replacement can be provided in a
".procs" file. For those files that are modified, the original file is
saved as ".bak" .

By default the script output shows which files the compiler is run on,
what compiler output is not recognized (and hence ignored) by the script,
which files are being modified, and what unexpected situations occur.
The latter are indicated as "errors".

The present limitations or the script are:
- the following idioms are flagged as "unimplemented" errors but not converted:
   - uses of 'iter' as an identifier,
   - 'return EXPR' in iterators;
- source code may not be converted when:
   - there are compilation errors,
   - 'def' occurs on a different line than the function name,
   - 'def' is preceeded or followed by non-white space;
- the module code is not handled
  (but the Chapel distribution contains already-converted module code).

(More:)

The recommended mode is: run this script on the code base, then
run the compiler again and follow up on its diagnostic messages.

Additional functionality that is desired:
- exclude certain files (e.g. that take very long time to compile)
- specify compiler options for specific files


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18218 3a8e244f-b0f2-452b-bcba-4c88e055c3ca