Improved error reporting for parallel testing, by allowing errors reported by paratest.client (as opposed to errors in the tests) be visible when it is invoked by paratest.server.
Note: if paratest.client returns an error, paratest.server will loop
forever. (It may or may not be worth the effort to fix that.)

Factored out error-reporting code in paratest.client.
Added minor comments.

"ssh -x" : without -x, paratest.client was occasionally giving me
errors like "xauth: failed to lock ~/.Xauthority", so -x makes
it easier on everyone.

Added the option -show-all-errors so that parallel testing
reports all errors as they are encountered - sometimes I find that
convenient. (I suspect with this option we will be getting
duplicates of error messages, but it is easier just to let it be.)

Added a test for PARAHALT to ensure paratest.server terminates
(in certain cases) when this file is created.

BTW if PARAHALT is created on a different host than paratest.server
is running on but they share the files via NFS, it may take
some time for the file to propagate to the other host,
during which time paratest.server will keep running.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18220 3a8e244f-b0f2-452b-bcba-4c88e055c3ca