Improve floating point literal support
This commit adds support for floating point literals of the form
1.e...  Previously we only supported 1e... and 1.0e...

Our traditional rationale for not supporting this form is that it is
ambiguous with the definition of a paren-less method named 'e' on an
integer.  Having ignored the topic for a few years, failure to support
C-style floating point literals seems increasingly regrettable, so
this commit essentially breaks the ambiguity by favoring the floating
point literal.  Note that it is still possible to define the
paren-less method on an integer, you just have to use whitespace to
make it clear that it's not a floating point literal.  Given the
unlikelihood/arguably poor idea of defining a no-paren method named
'e' (or 'E') on integers, this seems like the right way to break this
ambiguity, particularly since no functionality is lost if whitespace
is applied.

I think the only remaining case that we don't support is "1." due to
the fact that doing it in the naive way causes "1..n" to be
interpreted as "1. + "." + "n".  As a background topic I'm refreshing
my memory as to what it would take to break this ambiguity in favor of
the ".." operator.

Committed a test that uses the new format along with defining a method
e on integers.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18223 3a8e244f-b0f2-452b-bcba-4c88e055c3ca