This commit adds a new test, radix sort, to the studies test directory.
This commit adds a radix sort test. In that the code is mostly serial, left to
do with this is capturing parallelism with an external call to int_fetch_add
or something similar. The radix sort on F is generated using a random stream
with a seed to pass nightly testing, but for real testing, the user should
not use a seed. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18224 3a8e244f-b0f2-452b-bcba-4c88e055c3ca