This commit fully removes any real output during work in test.
This commit fully removes any real output during work in
child_remains_after_parent.chpl. Was hoping the setup to the loop would take
long enough so a period wouldn't sneak in before the parent printed to the
console. Now, it just writes a blank character (to take some time).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18225 3a8e244f-b0f2-452b-bcba-4c88e055c3ca