Minor cleanup to dodge intel warning
This is a minor change to some pointer-bumping code which had used a
pointer set just before the beginning of a string to make the body of
the loop more consistent by avoiding an initial boundary condition.
The recent Intel compilers have started complaining about this which
has broken our build, so this change avoids taking that approach.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18235 3a8e244f-b0f2-452b-bcba-4c88e055c3ca