Redesign and reorganization of the communication layer
This commit implements the new comm layer interface, but also includes
other significant changes and improvements that were required/nice.

Part 1: Comm layer interface change

There are two significant changes to the comm layer interface.  First,
the get and put routines now have an length parameter, in preparation
for bulk communication (currently we pass in length=1 for all cases).
In addition, all routines that send or receive data (get, put, fork,
broadcast) now have a "type index" parameter that corresponds to the
type of the element (along with the element size).  This integer is
used to index into compiler generated arrays that describe the
format/size of data being sent.  This value is used primarily for
heterogeneous compilation.  As a result, the --heterogeneous compiler
flag will work when compiling with any comm layer (not just PVM); the
generated code is only increased by the size of the type tables.

The changes required are to both the compiler interface (comm macros)
as well as the comm layer routines.  Specifically, in the compiler,
code generation to the macros was changed; in the runtime, the macro
definitions as well as the routines themselves were changed (for all
comm layers).

Additional changes made:

- Remove special cases of codegen (i.e., directly printing a string
  rather than using the symbol).  Specifically, dtObject (object base
  type) and (newly added) dtLocale (locale class).  Made an attempt at
  removing chpl__class_id chpl_cid field in the object class, but
  there were too many assumptions made about the object class during
  compilation.  It's not clear to me that not-special-casing the
  chpl_cid field is the right answer either, because non-codegen
  passes may have to deal with special cases for the object class.

- Other codegen clean up (replace fprintf() with gen(), make gen()
  more flexible with more accepted format characters, etc.)

- Better documentation of the "special" types in the compiler such as
  dIteratorRecord (internal), dtArray (required to be defined by the
  modules), and dtObject.  Added error checking for those that are
  required to be defined by the modules.  Changed dtEnumerated to
  dtAnyEnumerated for consistency.

- Add *all* types to the type index description arrays for
  heterogeneous compilation.  Previously these were being added as
  needed, but that was messy and error-prone (there were at least a
  couple of places where we were not adding a particular type and
  getting lucky).

- Added a flag on PrimitiveType that indicated an internal type.
  These should not be codegen'ed.  This was a way to avoid putting
  internal types into the type tables, but also it seemed good to
  distinguish these from Chapel primitive types.

- Remove #define CHPL_HETEROGENEOUS in the generated code as well as
  the runtime (PVM comm layer).  Added a global var called
  chpl_heterogeneous in the generated code that controls the code
  path.  As a result make/comm/Makefile.pvm and the directory itself
  are removed.

- Added the rest of the Chapel primitive types to the chplType enum
  (in runtime/include/chpltypes.h).  These are not yet handled in the
  PVM comm layer.

- Add macros for non-blocking puts in the runtime


Part 2: Runtime file (and function) naming and documentation

Greg, Brad, and I agreed upon a new file and function naming scheme.
The convention is described in runtime/README.  In this commit, I am
trying it out for the communication routines.  Assuming all goes well,
we should employ this convention throughout the runtime.

In addition, I pulled out the compiler macros into a separate header
file (chpl-comm-compiler-macros.h) and documented the use of the comm
interface routines.

As part of this renaming, we should get rid of all "platform"
subdirectories in runtime/include as possible.  In this commit,
runtime/xt-cle is removed.


REMAINING QUESTIONS/TODOS:

- Should runtime routines be able to use the compiler macros?
  Currently, the only use is in chpl_wide_string_copy() in
  runtime/src/chpltypes.c which calls CHPL_COMM_WIDE_GET_STRING().

- How do we get the type index info (for primitive types) to runtime
  routines.  Currently, the routines in chplmemcheck.c will not work
  for PVM true heterogeneous (that was true before this commit).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@18321 3a8e244f-b0f2-452b-bcba-4c88e055c3ca